+++ src/event/update.js
@@ -37,6 +37,26 @@ extends Event
         return this.message.id;
     }
 
+    switchedBranch() {
+
+        const placeEntry = this.diff.find(entry => entry.key === 'place');
+
+        if (placeEntry === undefined) {
+            return false;
+        }
+
+        if (placeEntry.oldValue.indexOf(':') !== -1
+            && placeEntry.newValue.indexOf(':') !== -1)
+        {
+            const levelBefore = placeEntry.oldValue.split(':');
+            const levelAfter  = placeEntry.newValue.split(':');
+
+            return levelBefore[0] !== levelAfter[0];
+        }
+
+        return placeEntry.oldValue !== placeEntry.newValue;
+    }
+
 }
 
 module.exports = Update;